# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Defensive network intelligence API for FadalWatch Network Guardian
 * OpenAPI spec version: 0.2.0
 */
import type { AlertSeverity } from './alertSeverity';
import type { AlertState } from './alertState';

export interface Alert {
  id: string;
  title: string;
  reason: string;
  severity: AlertSeverity;
  state: AlertState;
  assetId: string;
  createdAt: Date;
  recommendation: string;
  evidence?: string[];
}