# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Defensive network intelligence API for FadalWatch Network Guardian
 * OpenAPI spec version: 0.2.0
 */
import type { AssetBaselineState } from './assetBaselineState';
import type { AssetStatus } from './assetStatus';
import type { AssetType } from './assetType';

export interface Asset {
  id: string;
  address: string;
  label: string;
  type: AssetType;
  status: AssetStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  exposure: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  risk: number;
  lastSeen: Date;
  firstSeen?: Date;
  /** @minimum 0 */
  services: number;
  baselineState: AssetBaselineState;
}