# lib/api-zod/src/generated/types/assetDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Defensive network intelligence API for FadalWatch Network Guardian
 * OpenAPI spec version: 0.2.0
 */
import type { Activity } from './activity';
import type { Asset } from './asset';
import type { Service } from './service';

export type AssetDetail = Asset & {
  zone: string;
  servicesList: Service[];
  history: Activity[];
};