# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Defensive network intelligence API for FadalWatch Network Guardian
 * OpenAPI spec version: 0.2.0
 */

export interface HealthStatus {
  status: string;
}