# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Defensive network intelligence API for FadalWatch Network Guardian
 * OpenAPI spec version: 0.2.0
 */
import type { ServiceState } from './serviceState';

export interface Service {
  name: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  port: number;
  state: ServiceState;
  lastObserved: Date;
  /** @nullable */
  note?: string | null;
}